-fix warnings
git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@303 e54a6e83-5905-42d5-8d5c-058d10e6a962

+++ src/gui/imageinfo.cpp
@@ -95,7 +95,7 @@ int CImageInfo::exec(CMenuTarget* parent, const std::string &)
 	while (1)
 	{
 		neutrino_msg_data_t data;
-		unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd_MS(100);
+		uint64_t timeoutEnd = CRCInput::calcTimeoutEnd_MS(100);
 		g_RCInput->getMsgAbsoluteTimeout( &msg, &data, &timeoutEnd );
 
 		if (msg <= CRCInput::RC_MaxRC)
